refactor(native): CURL_READFUNC_ABORT 声明由 long 改为 int
对齐 curl.h 里 `#define CURL_READFUNC_ABORT 0x10000000` 的 int 字面量语义。
使用 int 而非 long 后,read callback 返回路径上的 `(UIntPtr)CURL_READFUNC_ABORT`
与其它 `(UIntPtr)int` 形式统一 —— 非负 int 转 UIntPtr 在 32/64 位平台都恒定安全,
消除原先 "long→UIntPtr 在 32 位平台会不会截断" 的类型歧义(实际上常量值 2^28
远小于 2^31, 即使按 long 截断也安全, 但声明成 int 更能让读代码的人一眼看清)。

+++ Packages/com.basecity.curl-unity/Runtime/Native/CurlNative.cs
@@ -82,7 +82,10 @@ namespace CurlUnity.Native
         public const int CURLOPT_INFILESIZE_LARGE = 30115;
 
         // READFUNCTION sentinel return values
-        public const long CURL_READFUNC_ABORT = 0x10000000;
+        // 对齐 curl.h 宏 #define CURL_READFUNC_ABORT 0x10000000 的 int 字面量语义。
+        // 使用 int 而非 long,避免跨平台 (UIntPtr)long 强转时的类型歧义 —— 非负 int 转
+        // UIntPtr 在 32/64 位平台都恒定安全,不依赖位宽讨论。
+        public const int CURL_READFUNC_ABORT = 0x10000000;
         public const int CURLOPT_SSL_OPTIONS = 216;
         public const long CURLSSLOPT_NATIVE_CA = 1 << 4;  // Use OS native CA store (Windows CryptoAPI)
